refactor: (#170) QuickStart 도메인의 dto 구조를 controller 사용 여부에 따라 재구성한다.
