+++ src/main/java/spring/backend/quickstart/domain/service/CreateQuickStartService.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import spring.backend.quickstart.domain.entity.QuickStart;
 import spring.backend.quickstart.domain.repository.QuickStartRepository;
-import spring.backend.quickstart.dto.request.QuickStartRequest;
+import spring.backend.quickstart.presentation.dto.request.QuickStartRequest;
 import spring.backend.quickstart.exception.QuickStartErrorCode;
 import spring.backend.core.util.TimeUtil;
 import spring.backend.member.domain.entity.Member;
