+++ src/main/java/spring/backend/quickstart/presentation/ReadQuickStartsController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 import spring.backend.quickstart.application.ReadQuickStartsService;
-import spring.backend.quickstart.dto.response.QuickStartsResponse;
+import spring.backend.quickstart.presentation.dto.response.QuickStartsResponse;
 import spring.backend.quickstart.presentation.swagger.ReadQuickStartsSwagger;
 import spring.backend.core.configuration.argumentresolver.AuthorizedMember;
 import spring.backend.core.configuration.interceptor.Authorization;
