+++ src/main/java/spring/backend/quickstart/presentation/UpdateQuickStartController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import spring.backend.quickstart.domain.service.UpdateQuickStartService;
-import spring.backend.quickstart.dto.request.QuickStartRequest;
+import spring.backend.quickstart.presentation.dto.request.QuickStartRequest;
 import spring.backend.quickstart.presentation.swagger.UpdateQuickStartSwagger;
 import spring.backend.core.configuration.argumentresolver.AuthorizedMember;
 import spring.backend.core.configuration.interceptor.Authorization;
