+++ src/main/java/spring/backend/quickstart/presentation/dto/request/QuickStartRequest.java
@@ -1,4 +1,4 @@
-package spring.backend.quickstart.dto.request;
+package spring.backend.quickstart.presentation.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.*;
